requirements.txt
numpy
scipy
scikit-learn
psutil
reportlab
pyyaml
requests
pytest
